Modified port selection to allow for named pipes
Changed the listen function not to just check for ports as numbers but
to allow for named pipes which appear as strings to be used.  In the
scenario that the swagger-server is used within the azure webapp the
swagger specification host entry can not be used to determine the port
number.  Specifically the scenario that the changes address is the
following:
server.js
var port = process.env.port || 1337;
app.listen(port, function () {
swagger spec
no host entry